fix: refactor of mrnu button of post
